Small fixes + inp. validation

+++ CampusCrib/BookingDetailsPage.xaml.cs
@@ -21,16 +21,34 @@ namespace CampusCrib
 
             bookingdetailspagevm = new BookingDetailsPageVM();
             BindingContext = bookingdetailspagevm;
+
+            // For setting the minimum start date to today, minimum end date to tomorrow
+            DateTime today = DateTime.Now;
+            pickerStartDate.MinimumDate = today;
+            var tomorrow = today.AddDays(1);
+            pickerEndDate.MinimumDate = tomorrow;
         }
 
-        private void btnUpdate_Clicked(object sender, EventArgs e)
+        private async void btnUpdate_Clicked(object sender, EventArgs e)
         {
-            bookingdetailspagevm.UpdateBooking();
+            var ans = await DisplayAlert("Update Booking", "Are you sure your updated details are correct?\n\nOld details will be lost."
+              , "Yes", "No");
+            if (ans)
+            {
+                bookingdetailspagevm.UpdateBooking();
+            }
+         
         }
 
-        private void btnDelete_Clicked(object sender, EventArgs e)
+        private async void btnDelete_Clicked(object sender, EventArgs e)
         {
-            bookingdetailspagevm.DeleteBooking();
+            var ans = await DisplayAlert("Delete Booking", "Are you sure you want to delete this booking?\n\nData will be lost forever."
+               , "Yes", "No");
+            if (ans)
+            {
+               bookingdetailspagevm.DeleteBooking();
+            }
+     
         }
     }
 }
